tests/wpt/tests/container-timing/tentative/nested-containertiming-nesting-change-to-invalid.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Container Timing: two nested containertiming nodes, parent switching from transparent to invalid nesting policy, with a child img inside of the inner</title>
<body>
<style>
body {
  margin: 0;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/container-timing/resources/container-timing-helpers.js"></script>
<script src="/element-timing/resources/element-timing-helpers.js"></script>
<script>
  let beforeRender;
  async_test(function (t) {
    assert_implements(window.PerformanceContainerTiming, "PerformanceContainerTiming is not implemented");
    const observer = new PerformanceObserver(
      t.step_func_done(function(entryList) {
        assert_equals(entryList.getEntries().length, 1, 'one entry expected for the image, for the inner containertiming');
        const entry = entryList.getEntries()[0];
        checkContainerEntry(entry, 'div2_ct', 'img_id', beforeRender)
        checkRect(entry, [0, 100, 0, 100])
        checkContainerSize(entry, 10000);
      })
    );
    observer.observe({entryTypes: ['container']});

    // Add a div that is the container timing root
    const div1 = document.createElement('div');
    div1.setAttribute('containertiming', 'div1_ct');
    div1.setAttribute('containertiming-nesting', 'transparent');
    document.body.appendChild(div1);
    div1.setAttribute('containertiming-nesting', 'invalid');

    // Add another div, child of the first, that is also a container root
    const div2 = document.createElement('div');
    div2.setAttribute('containertiming', 'div2_ct');
    div1.appendChild(div2);

    // Add image of width equal to 100 and height equal to 100.
    img = document.createElement('img');
    img.src = '/container-timing/resources/square100.png';
    img.setAttribute('id', 'img_id');
    div2.appendChild(img);

    beforeRender = performance.now();
  }, 'A parent containertiming root switches from transparent nesting policy to invalid, generating only a paint for the inner container timing.');
</script>

</body>
